# issue-7599: Optimize common cases on read path