--- a/lib/storage/storage_test.go
+++ b/lib/storage/storage_test.go
@@ -3339,7 +3339,7 @@ func TestStorageSearchMetricNamesWithoutPerDayIndex(t *testing.T) {
 	)
 	rng := rand.New(rand.NewSource(1))
 	opts := testStorageSearchWithoutPerDayIndexOptions{
-		wantEmpty:        []string{},
+		wantEmpty:        []string(nil),
 		wantPerTimeRange: make(map[TimeRange]any),
 		wantAll:          []string{},
 	}
@@ -3557,7 +3557,7 @@ func TestStorageSearchGraphitePathsWithoutPerDayIndex(t *testing.T) {
 	)
 	rng := rand.New(rand.NewSource(1))
 	opts := testStorageSearchWithoutPerDayIndexOptions{
-		wantEmpty:        []string{},
+		wantEmpty:        []string(nil),
 		wantPerTimeRange: make(map[TimeRange]any),
 		wantAll:          []string{},
 	}