@hagen1778 2025-05-14
https://github.com/VictoriaMetrics/VictoriaMetrics/issues/7599